Refactor to use a Board component
