+++ src/Board.css
@@ -1,0 +1,16 @@
+.Board {
+  display: grid;
+  grid-template-columns: repeat(3, 100px);
+  grid-template-rows: repeat(3, 100px);
+}
+
+.Tile {
+  border: 1px solid black;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+.Tile > img {
+  width: 100px;
+}
