+++ src/Game.css
@@ -3,20 +3,3 @@
   grid-template-columns: 350px 150px;
   margin-left: 50px;
 }
-
-.Board {
-  display: grid;
-  grid-template-columns: repeat(3, 100px);
-  grid-template-rows: repeat(3, 100px);
-}
-
-.Tile {
-  border: 1px solid black;
-  display: flex;
-  align-items: center;
-  justify-content: center;
-}
-
-.Tile > img {
-  width: 100px;
-}
